Sort: Selection and heap sort implemented, error handling improved
In case an error has occurred, the XT defined errno will be returned (in
this case, only XT_ENOMEM or maybe XT_EINVAL can be returned).

We have to write a couple of tests to verify selection and heap sort for
generic types. The insertion sort has been verified to work properly.