Sort: Radix sorting (generic types not included)
Generic types cannot be included at the moment without breaking the API
call because it must provide a function pointer for the indexing process
in order to fill the buckets correctly.
The second phase of radix sort can be achieved with either counting or
bucket sort (we use the latter). In both cases, we need to be able to
compute the bucket index. At the moment, sorting with generic types will
fail for radix sort with XT_EINVAL.