File: New API created
Finally a new handy file API has arrived! It is not yet complete.
Some other functions including mmap should be added in the future.

A little thing about the code, it has been tested, but not
to the extreme. The Windows code very checking the file size is
very awkward. It is just downright impossible to get it working
through kernel calls. This resulted in the classical way of using
fseek and ftell to retrieve the file size.