os: Changed xtGetProcessCount() on Linux a bit
The function now calls ps with different arguments.
The way that the buffer length is specfied has also changed.
Unfortunately the function still returns the wrong amount of
processes. I do not know of a sane way to fix this problem.

Using sysinfo() instead of executing some commands is no solution
to this problem. sysinfo() returns the amount of threads in total
or something instead of the amount of processes. Although it's
documentation for Linux clearly states otherwise...