List: Fixed a fatal bug and created a demo
When creating a list, specifying zero as size resulted in a segfault
because we forgot to include to check the capacity argument.

The header has been edited to have a maximum line length of 72 chars.

A demo has been created which tests the list. I think that the demo
is quite alright. The list seems to work properly now.