Hashmap: Resizing, removing and hashmap demo added
A hashmap will automatically resize if growing is enabled. If you remove
an element you can specify whether the key and/or value are also freed.

The iterator control functions are separated from the hashmap which
makes it easier to do a foreach loop using an external iterator instance
to make sure the hashmap's internal state does not change.

The demo performs very strict tests that ensure all data is kept intact
and correctly copied if the hashmap is resized.

Also removed a cast for malloc in xtHashmapAdd because it is redundant.