Thread: Mutexes now have destruction safeguards
It seems that Linux and Windows both have safeguards already in place.
This means that we don't have to add anything to the code!
On Linux it "is" possible that when the creation of pthread_mutexattr_t
fails, the mutex would be in an undefined state and cause a segfault
when destroying it. However the glibc source code shows that
pthread_mutexattr_init() cannot fail so it should be no problem.

The threads will get no destruction safeguards since it doesn't really
make any sense. You may be able to join the thread twice then but once
it is running, you cannot just terminate it anyways.