Thread: Fatal bugs fixed
Bugs which caused segfaults have been discovered and fixed!

On Windows it turned out that the suspendMutex was not initialized.
This resulted in a segfault when the thread was going to suspend.

The cleanup code on Windows was also wrong. The xtThreadJoin() function
has been altered to return void. Returning a boolean makes no sense
since you would probably get a segfault anyways when calling the
function a second time. (I know I do)