Time: Redesigned API, fixed Windows warning
From now on there are no seperate functions for every clock type.
Instead a single function is used which receives the desired clock
type to be worked with. The API is not complete yet.

Fixed warning between signed and unsigned ints. Documentation and the
demos are still to be updated.