Time: xtTimestamp API added
xtTimestamps can be formatted using a custom specified format. You can
also compute the difference between two timestamps. The string demo also
includes the appropriate xtTimestamp tests.

The supported format arguments are not documented in the header yet. The
demo should be descriptive enough to understand how the basics work
though.

We have to consider whether we want to support printf like precision,
flags and other optional arguments in the format string. This is very
time-consuming so this has been left out for now.