String: Added xtsnprintf()
Simple but working printf family added. This is also going to resolve
many issues with the broken printf family on windoze (e.g. safeguard
termination).

Linux' printf family is much better, but still missing things like %I64
for PRId64 and others. Arguments specified as I(8|16|32|64)[diuoxX] are
interpreted as (u)intx_t arguments and substituted with the
corresponding PRIdx string. Other inttypes such as leastx, fastx, max
and ptr are not interpreted yet.

The flag characters, field width, precision type and precision are
recognized for each provided argument in the format string. The format
argument type `I' is already used as a glibc extension, but we don't
need that so we can safely override it.

Another special argument that gets interpreted is %M which is similar
to the glibc extension %m for strerror(errno) but %M is the same as
xtGetErrorStr(errno).

xtsnprintf fails if the format string has to be resized when it doesn't
fit. We may want to implement auto-resizing if everything is working
properly.

Debug statements in xtsnprintf can be enabled by defining
XT_PRINTF_DEBUG. Windows' _xt_vsnprintf now also ensures proper NULL
termination.