Proc: xtProcGetMemoryInfo() has been added
An exciting new function has been added to the proc API! Too bad it
doesn't work very well under fuckin' Windows for a change. The 'swap'
and 'peak' values that Windows gives us are just complete bullshit. I
checked it tons of times in Process Explorer and Process Hacker and it
doesn't match up. That's why it's better just to set them to zero
instead.

The README has been updated because there is a new linker dependency.