Hash: Removed ascii string buffer
The ascii string buffer is not necessary since you can easily output
the hash to hex with xtFormatHex.

The demos are broken because xtFormatHex() requires a separator but
using one in our case doesn't make sense. That needs to be changed
pronto.

You can also now just retrieve the hash size for the current algorithm
easily with context.hashSizeInBytes.