File: Change xtFileIteratorStart behavior to Linux behavior
Added ERROR_NO_MORE_FILES to the error list so that if there are no
more files in a file iteration the proper XT error code is returned.

xtFileIteratorStart() on Windows now can e.g. fail to find the directory
on the first call just like on Linux. Previously Windows secretly did
nothing on the iterator start call. Right now the result to the first
file is cached and is returned on the first call to xtFileIteratorNext().