Demos: Added Socket connect loop demo
A basic connection test with a master and slave is simulated. It still
lacks proper synchronisation and resource handling.

Two threads are spawned where one thread connects to the other using the socket.
It does not send any data yet. Fixed some warnings as well.

Finally, let slave retry a couple of times